Prefer ReadUInt8 over ReadByte.
The former will throw when the end of the stream is reached, rather than requiring the caller to check for -1.

+++ OpenRA.Mods.Cnc/SpriteLoaders/ShpD2Loader.cs
@@ -21,7 +21,7 @@ namespace OpenRA.Mods.Cnc.SpriteLoaders
 	public class ShpD2Loader : ISpriteLoader
 	{
 		[Flags]
-		enum FormatFlags : int
+		enum FormatFlags : ushort
 		{
 			PaletteTable = 1,
 			NotLCWCompressed = 2,
